Remove unneeded StyledTextAreaModel specified to use StyledText<S> and rename Generic version back to original name
