+++ richtextfx/src/test/java/org/fxmisc/richtext/model/StyledTextAreaModelTest.java
@@ -15,7 +15,7 @@ public class StyledTextAreaModelTest {
 
         String text1 = "abc\r\ndef";
         String text2 = "A\r\nB\r\nC";
-        StyledTextAreaModel<Collection<String>, Collection<String>> model = new StyledTextAreaModel<>(
+        StyledTextAreaModel<Collection<String>, StyledText<Collection<String>>, Collection<String>> model = new StyledTextAreaModel<>(
                 Collections.<String>emptyList(),
                 Collections.<String>emptyList(),
                 segOps
@@ -36,7 +36,7 @@ public class StyledTextAreaModelTest {
 
         // set up area with some styled text content
         boolean initialStyle = false;
-        StyledTextAreaModel<String, Boolean> model = new StyledTextAreaModel<>(
+        StyledTextAreaModel<String, StyledText<Boolean>, Boolean> model = new StyledTextAreaModel<>(
                 "", initialStyle, new SimpleEditableStyledDocument<>("", initialStyle), segOps, true);
         model.replaceText("testtest");
         model.setStyle(0, 8, true);
